+++ test/controller.js
@@ -117,6 +117,57 @@ describe('Controller', function(){
       Leap.Controller._pluginFactories = {}
     });
 
+    it('should list plugins being used', function(){
+      Leap.Controller._pluginFactories = {}
+      Leap.Controller.plugin('testPlugin', fakePluginFactory());
+
+      var scope = {color: 'blue'};
+      var plugins = fakeController()
+        .use('testPlugin', scope)
+        .plugins;
+
+      assert.equal(plugins.testPlugin, scope);
+      Leap.Controller._pluginFactories = {}
+    });
+
+    it('should merge options when re-using', function(){
+      Leap.Controller._pluginFactories = {}
+      Leap.Controller.plugin('testPlugin', fakePluginFactory());
+
+      var scope = {color: 'blue'};
+      var plugins = fakeController()
+        .use('testPlugin', scope)
+        .use('testPlugin', {color: 'red'})
+        .plugins;
+
+      assert.equal(plugins.testPlugin, scope);
+      assert.equal(scope.color, 'red');
+      Leap.Controller._pluginFactories = {}
+    });
+
+    it('should list plugins being used after stopping using a plugin', function(){
+      Leap.Controller._pluginFactories = {}
+      Leap.Controller.plugin('testPlugin', fakePluginFactory());
+
+      var plugins = fakeController()
+        .use('testPlugin')
+        .stopUsing('testPlugin')
+        .plugins;
+
+      assert.equal(plugins.testPlugin, undefined);
+      Leap.Controller._pluginFactories = {}
+    });
+
+    describe('use', function () {
+      it('should accept plugin factories directly', function () {
+        var controller = fakeController()
+        controller.use(function (options) {
+          assert.equal(options.x, 2)
+          return {}
+        }, {x: 2})
+      });
+    });
+
     it('should decorate hands, fingers, and pointables', function(){
       Leap.Controller._pluginFactories = {}
       Leap.Controller.plugin('testPlugin', fakePluginFactory({