+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.4.2 - 2014-03-12
+ - Allow controller.use to be called idempotently
+ - Fix issue where requestAnimationFrame would not be used in Chrome
+ - Exclude node code from browser js
+
+##0.4.1 - 2014-02-11
+ - Allow controller.use to accept a factory directly (node support)
+
 ##0.4.0 - 2014-01-30
  - Started changelog
  - Added plugins